libra: Remove IMS packages
The IMS packages relate to the IP Multimedia Subsystem. We don't necessarily
need these for a stock LineageOS build; there is no demand for them; and
there isn't any way to test them.

N.B. This commit only removes IMS configuration and the .jar/.apk files.
Libraries and executables remain.

Removing with this commit. In future, if we want to reinstate, we can.

Change-Id: I2418ff7cf7a94f28c93548a0d5784f146275e913